+++ lib/view_component/base.rb
@@ -37,11 +37,8 @@ module ViewComponent
     RESERVED_PARAMETER = :content
     VC_INTERNAL_DEFAULT_FORMAT = :html
 
-    # For CSRF authenticity tokens in forms
-    delegate :form_authenticity_token, :protect_against_forgery?, :config, to: :helpers
-
-    # For Content Security Policy nonces
-    delegate :content_security_policy_nonce, to: :helpers
+    # For CSRF authenticity tokens in forms and Content Security Policy nonces
+    use_helpers :form_authenticity_token, :protect_against_forgery?, :config, :content_security_policy_nonce
 
     # Config option that strips trailing whitespace in templates before compiling them.
     class_attribute :__vc_strip_trailing_whitespace, instance_accessor: false, instance_predicate: false
@@ -235,6 +232,7 @@ module ViewComponent
     # @return [ActionView::Base]
     def helpers
       raise HelpersCalledBeforeRenderError if view_context.nil?
+      raise StrictHelperError if ViewComponent::Base.strict_helpers_enabled
       # Attempt to re-use the original view_context passed to the first
       # component rendered in the rendering pipeline. This prevents the
       # instantiation of a new view_context via `controller.view_context` which
@@ -251,11 +249,19 @@ module ViewComponent
         super
       rescue => e # rubocop:disable Style/RescueStandardError
         e.set_backtrace e.backtrace.tap(&:shift)
-        raise e, <<~MESSAGE.chomp if view_context && e.is_a?(NameError) && helpers.respond_to?(method_name)
+        if ViewComponent::Base.strict_helpers_enabled
+          raise e, <<~MESSAGE.chomp if view_context && e.is_a?(NameError) && (__vc_original_view_context.respond_to?(method_name)|| controller.view_context.respond_to?(method_name))
           #{e.message}
 
-          You may be trying to call a method provided as a view helper. Did you mean `helpers.#{method_name}'?
-        MESSAGE
+            You may be trying to call a method provided as a view helper. To use it try decalring it using use_helpers :#{method_name}'?
+          MESSAGE
+        else
+          raise e, <<~MESSAGE.chomp if view_context && e.is_a?(NameError) && helpers.respond_to?(method_name)
+            #{e.message}
+
+            You may be trying to call a method provided as a view helper. Did you mean `helpers.#{method_name}'?
+          MESSAGE
+        end
 
         raise
       end