+++ test/sandbox/test/rendering_test.rb
@@ -184,6 +184,20 @@ class RenderingTest < ViewComponent::TestCase
     ActionController::Base.allow_forgery_protection = old_value
   end
 
+  def test_renders_button_to_component_with_strict_helpers
+    with_strict_helpers_config(true) do
+      old_value = ActionController::Base.allow_forgery_protection
+      ActionController::Base.allow_forgery_protection = true
+
+      render_inline(ButtonToComponent.new) { "foo" }
+
+      assert_selector("form[class='button_to'][action='/'][method='post']")
+      assert_selector("input[type='hidden'][name='authenticity_token']", visible: false)
+
+      ActionController::Base.allow_forgery_protection = old_value
+    end
+  end
+
   def test_renders_component_with_variant
     with_variant :phone do
       render_inline(VariantsComponent.new)
@@ -340,6 +354,22 @@ class RenderingTest < ViewComponent::TestCase
 
     component.config.asset_host = "assets.example.com"
     assert_match(%r{http://assets.example.com/assets/application-\w+.css}, render_inline(component).text)
+  end
+
+  def test_renders_component_with_asset_url_with_strict_helpers
+    with_strict_helpers_config(true) do
+      component = AssetComponent.new
+      assert_match(%r{http://assets.example.com/assets/application-\w+.css}, render_inline(component).text)
+
+      component.config.asset_host = nil
+      assert_match(%r{/assets/application-\w+.css}, render_inline(component).text)
+
+      component.config.asset_host = "http://assets.example.com"
+      assert_match(%r{http://assets.example.com/assets/application-\w+.css}, render_inline(component).text)
+
+      component.config.asset_host = "assets.example.com"
+      assert_match(%r{http://assets.example.com/assets/application-\w+.css}, render_inline(component).text)
+    end
   end
 
   def test_template_changes_are_not_reflected_if_cache_is_not_cleared
@@ -776,6 +806,14 @@ class RenderingTest < ViewComponent::TestCase
     assert_text("http://assets.example.com")
   end
 
+  def test_renders_component_using_rails_config_with_strict_helpers
+    with_strict_helpers_config(true) do
+      render_inline(RailsConfigComponent.new)
+
+      assert_text("http://assets.example.com")
+    end
+  end
+
   def test_inherited_component_inherits_template
     render_inline(InheritedTemplateComponent.new)
 
@@ -1128,6 +1166,14 @@ class RenderingTest < ViewComponent::TestCase
     assert_selector("script", text: "\n//<![CDATA[\n  \"alert('hello')\"\n\n//]]>\n", visible: :hidden)
   end
 
+  def test_content_security_policy_nonce_with_strict_helpers
+    with_strict_helpers_config(true) do
+      render_inline(ContentSecurityPolicyNonceComponent.new)
+
+      assert_selector("script", text: "\n//<![CDATA[\n  \"alert('hello')\"\n\n//]]>\n", visible: :hidden)
+    end
+  end
+
   def test_use_helper
     render_inline(UseHelpersComponent.new)
     assert_selector ".helper__message", text: "Hello helper method"
@@ -1218,6 +1264,14 @@ class RenderingTest < ViewComponent::TestCase
       render_inline(MultipleFormatsComponent.new)
 
       assert_equal(rendered_json["hello"], "world")
+    end
+  end
+
+  def test_strict_helpers
+    with_strict_helpers_config(true) do
+      assert_raises ViewComponent::StrictHelperError do
+        render_inline(HelpersProxyComponent.new)
+      end
     end
   end
 end