+++ test/test_helper.rb
@@ -176,6 +176,15 @@ end
 def with_compiler_development_mode(mode)
   previous_mode = ViewComponent::Compiler.development_mode
   ViewComponent::Compiler.development_mode = mode
+end
+
+def with_strict_helpers_config(enabled, &block)
+  with_config_option(:strict_helpers_enabled, enabled, &block)
+end
+
+def with_compiler_mode(mode)
+  previous_mode = ViewComponent::Compiler.mode
+  ViewComponent::Compiler.mode = mode
   yield
 ensure
   ViewComponent::Compiler.development_mode = previous_mode